requirements.txt
pandas
numpy
scikit-learn
matplotlib
PyYAML
dvc